cleans reorganizes code to better accomodate more modes, and improve perf
